move `ScreenBaseClass` to `lib/` dir
