+++ x-pack/plugins/security_solution/scripts/endpoint/agent_emulator/screens/lib/screen_base_class.ts
@@ -11,7 +11,7 @@ import type { WriteStream as TtyWriteStream } from 'tty';
 import { stdout, stdin } from 'node:process';
 import * as readline from 'node:readline';
 import { red, blue, green } from 'chalk';
-import { HORIZONTAL_LINE } from '../../common/constants';
+import { HORIZONTAL_LINE } from '../../../common/constants';
 
 class RenderedScreen {
   public statusPos: number = -1;
